Art resolvers
